# Conflicts:
#	CHANGELOG.md
#	install/changes/2.4.3.txt

+++ external/elFinder/php/elFinderVolumeDriver.class.php
@@ -282,6 +282,7 @@ abstract class elFinderVolumeDriver {
 			'rtf:text/rtf'                 => 'application/rtf',
 			'rtfd:text/rtfd'               => 'application/rtfd',
 			'ico:image/vnd.microsoft.icon' => 'image/x-icon',
+			'pxd:application/octet-stream' => 'image/x-pixlr-data',
 			'amr:application/octet-stream' => 'audio/amr',
 			'm4a:video/mp4'                => 'audio/mp4',
 			'oga:application/ogg'          => 'audio/ogg',
@@ -292,14 +293,21 @@ abstract class elFinderVolumeDriver {
 			'mpd:application/xml'          => 'application/dash+xml',
 			'xml:application/xml'          => 'text/xml',
 			'*:application/x-dosexec'      => 'application/x-executable',
-			'webp:application/octet-stream'=> 'image/webp',
 			'doc:application/vnd.ms-office'=> 'application/msword',
 			'xls:application/vnd.ms-office'=> 'application/vnd.ms-excel',
 			'ppt:application/vnd.ms-office'=> 'application/vnd.ms-powerpoint',
 			'yml:text/plain'               => 'text/x-yaml',
-			'docx:application/zip'         => 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-			'xlsx:application/zip'         => 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet xlsx',
-			'pptx:application/zip'         => 'application/vnd.openxmlformats-officedocument.presentationml.presentation'
+			'ai:application/pdf'           => 'application/postscript',
+			'cgm:text/plain'               => 'image/cgm',
+			'dxf:text/plain'               => 'image/vnd.dxf',
+			'hpgl:text/plain'              => 'application/vnd.hp-hpgl',
+			'igs:text/plain'               => 'model/iges',
+			'iges:text/plain'              => 'model/iges',
+			'plt:application/octet-stream' => 'application/plt',
+			'plt:text/plain'               => 'application/plt',
+			'sat:text/plain'               => 'application/sat',
+			'step:text/plain'              => 'application/step',
+			'stp:text/plain'               => 'application/step'
 		),
 		// An option to add MimeMap to the `mimeMap` option
 		// Array '[ext]:[detected mime type]' => '[normalized mime]'
@@ -4297,6 +4305,10 @@ abstract class elFinderVolumeDriver {
 			}
 			if ($type) {
 				if ($ext && preg_match('~^application/(?:octet-stream|(?:x-)?zip)~', $type)) {
+					// load default MIME table file "mime.types"
+					if (!elFinderVolumeDriver::$mimetypesLoaded) {
+						elFinderVolumeDriver::loadMimeTypes();
+					}
 					if (isset(elFinderVolumeDriver::$mimetypes[$ext])) {
 						$type = elFinderVolumeDriver::$mimetypes[$ext];
 					}
@@ -4612,7 +4624,7 @@ abstract class elFinderVolumeDriver {
 	 **/
 	protected function doSearch($path, $q, $mimes) {
 		$result = array();
-		$matchMethod = empty($this->doSearchCurrentQuery['matchMethod'])? 'stripos' : $this->doSearchCurrentQuery['matchMethod'];
+		$matchMethod = empty($this->doSearchCurrentQuery['matchMethod'])? 'searchMatchName' : $this->doSearchCurrentQuery['matchMethod'];
 		$timeout = $this->options['searchTimeout']? $this->searchStart + $this->options['searchTimeout'] : 0;
 		if ($timeout && $timeout < time()) {
 			$this->setError(elFinder::ERROR_SEARCH_TIMEOUT, $this->path($this->encode($path)));
@@ -5075,8 +5087,13 @@ abstract class elFinderVolumeDriver {
 				} catch (Exception $e) {}
 			}
 			if (! $result) {
-				file_exists($tmb) && unlink($tmb);
-				return false;
+				// fallback imgLib to GD
+				if (function_exists('gd_info') && ($s = getimagesize($tmb))) {
+					$this->imgLib = 'gd';
+				} else {
+					file_exists($tmb) && unlink($tmb);
+					return false;
+				}
 			}
 			$result = false;
 		}
@@ -5681,13 +5698,13 @@ abstract class elFinderVolumeDriver {
 	 * Execute shell command
 	 *
 	 * @param  string $command command line
-	 * @param  array $output stdout strings
-	 * @param array|int $return_var process exit code
-	 * @param  array $error_output stderr strings
+	 * @param  string $output stdout strings
+	 * @param  int $return_var process exit code
+	 * @param  string $error_output stderr strings
 	 * @return int exit code
 	 * @author Alexey Sukhotin
 	 */
-	protected function procExec($command , array &$output = null, &$return_var = -1, array &$error_output = null) {
+	protected function procExec($command , &$output = '', &$return_var = -1, &$error_output = '') {
 
 		static $allowed = null;
 		
@@ -5719,17 +5736,41 @@ abstract class elFinderVolumeDriver {
 		$process = proc_open($command, $descriptorspec, $pipes, null, null);
 
 		if (is_resource($process)) {
+			stream_set_blocking($pipes[1], 0);
+			stream_set_blocking($pipes[2], 0);
 
 			fclose($pipes[0]);
 
 			$tmpout = '';
 			$tmperr = '';
-
-			$error_output = stream_get_contents($pipes[2]);
-			$output = stream_get_contents($pipes[1]);
+			while (feof($pipes[1]) === false || feof($pipes[2]) === false) {
+				elFinder::extendTimeLimit();
+				$read = array($pipes[1], $pipes[2]);
+				$write = null;
+				$except = null;
+				$ret = stream_select($read, $write, $except, 1);
+				if ($ret === false) {
+					// error
+					break;
+				} else if ($ret === 0) {
+					// timeout
+					continue;
+				} else {
+					foreach ($read as $sock) {
+						if ($sock === $pipes[1]) {
+							$tmpout .= fread($sock, 4096);
+						} else if ($sock === $pipes[2]) {
+							$tmperr .= fread($sock, 4096);
+						}
+					}
+				}
+			}
 
 			fclose($pipes[1]);
 			fclose($pipes[2]);
+
+			$error_output = $tmpout;
+			$output = $tmperr;
 			$return_var = proc_close($process);
 
 		} else {
@@ -6008,6 +6049,19 @@ abstract class elFinderVolumeDriver {
 	}
 
 	/**
+	 * Default serach match method (name match)
+	 *
+	 * @param  String  $name  Item name
+	 * @param  String  $query Query word
+	 * @param  String  $path  Item path
+	 *
+	 * @return @return bool
+	 */
+	protected function searchMatchName($name , $query , $path) {
+		return $this->stripos($name , $query) !== false;
+	}
+
+	/**
 	 * Get server side available archivers
 	 * 
 	 * @param bool $use_cache
@@ -6054,12 +6108,12 @@ abstract class elFinderVolumeDriver {
 			unset($o);
 			$this->procExec(ELFINDER_ZIP_PATH . ' -h', $o, $c);
 			if ($c == 0) {
-				$arcs['create']['application/zip']  = array('cmd' => ELFINDER_ZIP_PATH, 'argc' => '-r9', 'ext' => 'zip');
+				$arcs['create']['application/zip']  = array('cmd' => ELFINDER_ZIP_PATH, 'argc' => '-r9 -q', 'ext' => 'zip');
 			}
 			unset($o);
 			$this->procExec(ELFINDER_UNZIP_PATH . ' --help', $o, $c);
 			if ($c == 0) {
-				$arcs['extract']['application/zip'] = array('cmd' => ELFINDER_UNZIP_PATH, 'argc' => '',  'ext' => 'zip', 'toSpec' => '-d ');
+				$arcs['extract']['application/zip'] = array('cmd' => ELFINDER_UNZIP_PATH, 'argc' => '-q',  'ext' => 'zip', 'toSpec' => '-d ');
 			}
 			unset($o);
 			$this->procExec(ELFINDER_RAR_PATH . ' --version', $o, $c);
